+++ CHANGES.md
@@ -2,6 +2,9 @@
 ##### Changes
 - Change "Aeroplane" to "Airplane"
 
+##### Fixes
+- Update Toolbar button tooltip when changing modes
+
 ## v0.2.0
 - Added option to invert pitch in aero mode.